requirements.txt
numpy==1.26.4
opencv-python==4.11.0.86
torch==2.2.0
transformers==4.39.3
pillow==11.1.0
flask==3.1.0
flask-cors==5.0.1
